requirements.txt
# Основные зависимости
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2
python-dotenv==1.0.0
httpx==0.25.2
cryptography==41.0.8
apscheduler==3.10.4

# Для разработки и тестирования
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
